refactor : customer 패키지 분리
